+++ Core/Src/Utils/LUT.c
@@ -1,4 +1,4 @@
-#include <stdbool.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #include "../Inc/Utils/LUT.h"
@@ -13,12 +13,12 @@ int point_compare(const void *a, const void *b) {
   return 0;
 }
 
-bool point_is_between(const point *min, const point *max, double in) {
+bool point_is_between(const point *min, const point *max, float in) {
   return min->input <= in && in <= max->input;
 }
 
 /* Returns true if the input can interpolate between the min and max. */
-bool can_interpolate(const point *min, const point *max, double in) {
+bool can_interpolate(const point *min, const point *max, float in) {
   // Either point is not defined, so can't interpolate
   if (min == NULL || max == NULL) {
     return false;
@@ -28,7 +28,7 @@ bool can_interpolate(const point *min, const point *max, double in) {
   return point_is_between(min, max, in);
 }
 
-bool table_init(table *table, unsigned long count) {
+bool table_init(table *table, uint16_t count) {
   if (table == NULL) {
     return false;
   }
@@ -42,7 +42,7 @@ bool table_init(table *table, unsigned long count) {
   table->count_ = count;
   table->added_ = 0;
   table->initialized_ = false;
-  for (unsigned long n = 0; n < count; ++n) {
+  for (uint16_t n = 0; n < count; ++n) {
     table->points_[n].input = 0;
     table->points_[n].output = 0;
   }
@@ -65,7 +65,7 @@ bool table_is_initialized(table *table) {
 
   bool sorted = true;
   const point *previous_point = &table->points_[0];
-  for (unsigned long n = 1; n < table->count_; ++n) {
+  for (uint16_t n = 1; n < table->count_; ++n) {
     const point *point = &table->points_[n];
     // Found a decrease, which violates table being sorted
     if (point->input < previous_point->input) {
@@ -76,7 +76,7 @@ bool table_is_initialized(table *table) {
   return table->initialized_;
 }
 
-bool table_add_reference_point(table *table, double in, double out) {
+bool table_add_reference_point(table *table, float in, float out) {
   if (table_is_initialized(table)) {
     return false;
   }
@@ -104,7 +104,7 @@ const point *table_max_point(table *table) {
   return &table->points_[table->count_ - 1];
 }
 
-bool table_can_sample(table *table, double in) {
+bool table_can_sample(table *table, float in) {
   const point *min = table_min_point(table);
   const point *max = table_max_point(table);
   return can_interpolate(min, max, in);
@@ -116,11 +116,11 @@ void table_sort(table *table) {
 
 /* Search for the index of the first reference point in the table that has a
  * greater input. */
-unsigned long table_search(const table *table, double in) {
+uint16_t table_search(const table *table, double in) {
   // TODO Implement a more efficient search algorithm; should not be necessary
   // for small tables
 
-  unsigned long n;
+  uint16_t n;
   for (n = 0; n < table->count_; n++) {
     if (table->points_[n].input > in) {
       break;
@@ -134,7 +134,7 @@ unsigned long table_search(const table *table, double in) {
   return n;
 }
 
-bool table_sample(table *table, double in, double *out) {
+bool table_sample(table *table, float in, float *out) {
   // This is where the constraint that the table contains at least two elements
   // is derived from
   const point *min = table_min_point(table);
@@ -157,12 +157,12 @@ bool table_sample(table *table, double in, double *out) {
   }
 
   // Search for the first point in the table that has a greater input
-  unsigned long n = table_search(table, in);
+  uint16_t n = table_search(table, in);
   point low = table->points_[n - 1];
   point high = table->points_[n];
 
   // Map the input value to a t-value [0, 1]
-  double t = (in - low.input) / (high.input - low.input);
+  float t = (in - low.input) / (high.input - low.input);
 
   *out = low.output + t * (high.output - low.output);
   return true;
